+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Version 4.3.2 So Far... [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - Fix duplicate TALENT entries in HeroItems compendium [#3980](https://github.com/dmdorman/hero6e-foundryvtt/issues/3980)
-- Automatons now have EGO characteristic. [#3986](https://github.com/dmdorman/hero6e-foundryvtt/issues/3986)
+- Automatons now have EGO characteristic. [#3989
+  ](https://github.com/dmdorman/hero6e-foundryvtt/issues/3989) [#3986](https://github.com/dmdorman/hero6e-foundryvtt/issues/3986) [#3768](https://github.com/dmdorman/hero6e-foundryvtt/issues/3768)
 
 ### Version 4.3.1 20260327